+++ mopl-api/src/main/java/io/mopl/api/ApiApplication.java
@@ -3,12 +3,10 @@ package io.mopl.api;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-import org.springframework.retry.annotation.EnableRetry;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
 @EnableScheduling
 @EnableJpaAuditing
-@EnableRetry
 @SpringBootApplication
 public class ApiApplication {
   public static void main(String[] args) {
