+++ mopl-api/src/main/java/io/mopl/api/auth/event/PasswordResetEvent.java
@@ -1,15 +1,15 @@
-package io.mopl.api.auth.event;
-
-import java.util.UUID;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-/** 비밀번호 초기화 완료 이벤트 */
-@Getter
-@AllArgsConstructor
-public class PasswordResetEvent {
-
-  private UUID userId;
-  private String email;
-  private String temporaryPassword;
-}
+// package io.mopl.api.auth.event;
+//
+// import java.util.UUID;
+// import lombok.AllArgsConstructor;
+// import lombok.Getter;
+//
+/// ** 비밀번호 초기화 완료 이벤트 */
+// @Getter
+// @AllArgsConstructor
+// public class PasswordResetEvent {
+//
+//  private UUID userId;
+//  private String email;
+//  private String temporaryPassword;
+// }
