+++ mopl-api/src/main/resources/application.yml
@@ -25,23 +25,23 @@ spring:
         spring.json.trusted.packages: io.mopl.core.event.*
   messages:
     basename: messages,ValidationMessages
-  mail:
-    host: smtp.gmail.com
-    port: 587
-    username: ${GMAIL_EMAIL}
-    password: ${GMAIL_APP_PASSWORD}
-    properties:
-      mail:
-        smtp:
-          auth: true
-          starttls:
-            enable: true
-            required: true
-          connectiontimeout: 5000
-          timeout: 5000
-          writetimeout: 5000
-          ssl:
-            trust: smtp.gmail.com
+#  mail:
+#    host: smtp.gmail.com
+#    port: 587
+#    username: ${GMAIL_EMAIL}
+#    password: ${GMAIL_APP_PASSWORD}
+#    properties:
+#      mail:
+#        smtp:
+#          auth: true
+#          starttls:
+#            enable: true
+#            required: true
+#          connectiontimeout: 5000
+#          timeout: 5000
+#          writetimeout: 5000
+#          ssl:
+#            trust: smtp.gmail.com
 security:
   cookie:
     same-site: ${COOKIE_SAME_SITE:Lax}
