+++ mopl-worker/build.gradle
@@ -18,4 +18,11 @@ dependencies {
     implementation 'software.amazon.awssdk:s3:2.40.17'
 
     testImplementation 'org.springframework.kafka:spring-kafka-test'
+
+    // Email 전송
+    implementation 'org.springframework.boot:spring-boot-starter-mail'
+
+    // 재시도 메커니즘
+    implementation 'org.springframework.retry:spring-retry:2.0.5'
+    implementation 'org.springframework:spring-aspects'
 }
