+++ mopl-worker/src/main/resources/application.yml
@@ -28,6 +28,23 @@ spring:
       ack-mode: manual
   jackson:
     time-zone: UTC
+  mail:
+    host: smtp.gmail.com
+    port: 587
+    username: ${GMAIL_EMAIL}
+    password: ${GMAIL_APP_PASSWORD}
+    properties:
+      mail:
+        smtp:
+          auth: true
+          starttls:
+            enable: true
+            required: true
+          connectiontimeout: 5000
+          timeout: 5000
+          writetimeout: 5000
+          ssl:
+            trust: smtp.gmail.com
 
 aws:
   s3:
